pyproject.toml
[project]
name = "ai-software-development"
version = "0.1.1"
description = "A modular, multi-agent system for end-to-end software development projects"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "deepagents>=0.0.3",
    "langgraph-cli[inmem]>=0.1.0",
    "tavily-python>=0.3.0",
    "PyYAML>=6.0",
    "ollama>=0.3.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
]
dev = [
    "black",
    "isort",
    "flake8",
    "mypy",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

## pytest configuration moved to pytest.ini to avoid duplication

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88
